chore: Optimize db query and data operations
    1.Optimize the dabase query flow.
    2.Add voice note folder & note operation utility.
    3.Fixed main window flash bug.(call moveToCenter before
    window being showed)

+++ deepin-voice-note/common/vnoteitem.h
@@ -20,9 +20,9 @@ public:
         VNT_Voice,
     };
 
-    qint64 noteId {INVALID_ID};
+    qint32 noteId {INVALID_ID};
     qint64 folderId {INVALID_ID};
-    qint8  noteType {VNOTE_TYPE::VNT_Text};
+    qint32 noteType {VNOTE_TYPE::VNT_Text};
 
     QString noteText;
     QString voicePath;
